chore: 장르 관련 API 함수 분리 및 모듈화 #45
