+++ src/api/movies.js
@@ -1,7 +1,6 @@
 import { get, REGION } from "./client";
 import { getBestTrailerUrl } from "./videos";
 import { getToday, getMonthsAgo } from "@utils/format";
-import { DEFAULT_SERVER_SORT } from "@utils/sort";
 
 // Hero Movies
 export async function fetchHeroMovies() {
@@ -106,34 +105,4 @@ export async function searchMovies(query, page = 1) {
     query,
     page,
   });
-}
-
-// Genres
-export async function fetchGenreMap() {
-  const data = await get("/genre/movie/list", { language: "ko" });
-  return Object.fromEntries((data.genres || []).map((g) => [g.id, g.name]));
-}
-
-// Discover By Genre
-export async function discoverByGenre(
-  genreId,
-  page = 1,
-  sortBy = DEFAULT_SERVER_SORT
-) {
-  const today = getToday();
-
-  const params = {
-    with_genres: genreId,
-    include_adult: false,
-    page,
-    sort_by: sortBy,
-    "release_date.lte": today, // 개봉된 영화만
-  };
-
-  // 평점순 정렬일 때는 최소 투표 수 제한
-  if (sortBy.startsWith("vote_average.")) {
-    params["vote_count.gte"] = 100;
-  }
-
-  return get("/discover/movie", params);
 }
